+++ condition.go
@@ -150,7 +150,7 @@ func ExpectIs(v any, kinds ...reflect.Kind) Condition {
 	return Condition{
 		result:   result,
 		trueMsg:  fmt.Sprintf("the value is %s", kindV),
-		falseMsg: fmt.Sprintf("the value is %s", kindV),
+		falseMsg: fmt.Sprintf("the value is %s, not %v", kindV, kinds),
 	}
 }
 
@@ -239,6 +239,66 @@ func ExpectError(err error, targets ...error) Condition {
 // targets.
 func ExpectErrorNot(err error, targets ...error) Condition {
 	return ExpectError(err, targets...).revert()
+}
+
+// ExpectIn returns a true Condition if the element is in the object which must
+// be an array, slice, string, or map.
+func ExpectIn(object any, element any) Condition {
+	AssertIs(object, reflect.Array, reflect.Slice, reflect.String, reflect.Map)
+
+	var objectV = reflect.ValueOf(object)
+	var elementV = reflect.ValueOf(element)
+
+	switch objectV.Kind() {
+	case reflect.Map:
+		AssertEqual(objectV.Type().Key(), elementV.Type())
+		var v = objectV.MapIndex(elementV)
+		return Condition{
+			result:   v != reflect.Value{},
+			trueMsg:  fmt.Sprintf("element %v existed in map", element),
+			falseMsg: fmt.Sprintf("element %v doesn't exist in map", element),
+		}
+	case reflect.Slice, reflect.Array:
+		AssertEqual(objectV.Type().Elem(), elementV.Type())
+		var cond = Condition{
+			result:   false,
+			trueMsg:  fmt.Sprintf("element %v existed in array", element),
+			falseMsg: fmt.Sprintf("element %v doesn't exist in array", element),
+		}
+		for i := 0; i < objectV.Len(); i++ {
+			if element == objectV.Index(i).Interface() {
+				cond.result = true
+				break
+			}
+		}
+		return cond
+	case reflect.String:
+		AssertIs(element, reflect.Int32, reflect.String)
+		var cond = Condition{
+			result:   false,
+			trueMsg:  "string contains the element",
+			falseMsg: "string doesn't contain the element",
+		}
+		switch elementV.Kind() {
+		case reflect.Int32:
+			cond.result = strings.ContainsRune(object.(string), element.(rune))
+		case reflect.String:
+			cond.result = strings.Contains(object.(string), element.(string))
+		}
+		return cond
+	}
+
+	return Condition{
+		result:   false,
+		trueMsg:  "an unknown error occurred",
+		falseMsg: "an unknown error occurred",
+	}
+}
+
+// ExpectNotIn returns a true Condition if the element is not in the object
+// which must be an array, slice, string, or map.
+func ExpectNotIn(object any, element any) Condition {
+	return ExpectIn(object, element).revert()
 }
 
 // ExpectTrue returns true if the the parameter is true.