+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Developing
+
+-   Add ExpectIn supports to check if an element is in object.
+
 # V0.0.1 (Sep 02, 2022)
 
 -   Migrated from the [origin project](https://github.com/xybor/xyplatform).